requirements.txt
torch>=2.0.0
torchvision>=0.15.0
python-chess>=1.999
numpy>=1.24.0
matplotlib>=3.6.0
tqdm>=4.64.0
tensorboard>=2.12.0
flask>=2.2.0
requests>=2.28.0
scipy>=1.10.0